requirements.txt
matplotlib
pyyaml
psutil
requests
svgwrite
plotly==5.13.1
numpy